Remove consent screen from Biometric KYC
