+++ Example/SmileID/BiometricKYC/IdInfoInputViewModel.swift
@@ -1,4 +1,5 @@
 import Combine
+import SmileID
 import SwiftUI
 import UIKit
